@@ -6,25 +6,15 @@ namespace EscapeFromTheWoods
 {
     public class DBMonkeyRecord
     {
-        public DBMonkeyRecord(int monkeyID, string monkeyName, int woodID, int seqNr, int treeID, int x, int y)
+        public DBMonkeyRecord(int monkeyID, string monkeyName, List<Tree> route)
         {
-            //this.recordID = recordID;
             this.monkeyID = monkeyID;
             this.monkeyName = monkeyName;
-            this.woodID = woodID;
-            this.seqNr = seqNr;
-            this.treeID = treeID;
-            this.x = x;
-            this.y = y;
+            this.route = route;
         }
 
-        public int recordID { get; set; }
         public int monkeyID { get; set; }
         public string monkeyName { get; set; }
-        public int woodID { get; set; }
-        public int seqNr { get; set; }
-        public int treeID { get; set; }
-        public int x { get; set; }
-        public int y { get; set; }
+        public List<Tree> route { get; set; }
     }
 }
