@@ -1,22 +1,32 @@
-﻿using System;
+﻿using EscapeFromTheWoods.Objects;
+using System;
 using System.Collections.Generic;
+using System.Linq;
 using System.Text;
+using System.Threading.Tasks;
 
 namespace EscapeFromTheWoods
 {
     public static class WoodBuilder
-    {        
-        public static Wood GetWood(int size,Map map,string path,DBwriter db)
+    {
+        public static Wood GetWood(int size, Map map, string path, DBwriter db)
         {
+            // Changed tree checking by List.contains to dictionary 
             Random r = new Random(100);
-            List<Tree> trees = new List<Tree>();
+            Dictionary<int, Tree> trees = new Dictionary<int, Tree>();
             int n = 0;
-            while(n<size)
+            int index;
+            while (n < size)
             {
-                Tree t = new Tree(IDgenerator.GetTreeID(),r.Next(map.xmin,map.xmax),r.Next(map.ymin,map.ymax));
-                if (!trees.Contains(t)) { trees.Add(t); n++; }
+                index = r.Next(size);
+                if (!trees.ContainsKey(index))
+                {
+                    n++;
+                    trees.Add(index, new Tree(IDgenerator.GetTreeID(), r.Next(map.xmin, map.xmax), r.Next(map.ymin, map.ymax)));
+                }
             }
-            Wood w = new Wood(IDgenerator.GetWoodID(),trees,map,path,db);
+
+            Wood w = new Wood(IDgenerator.GetWoodID(), trees.Values.ToList(), map, path, db);
             return w;
         }
     }
